bug fix to the rule evaluation order introduced when rangeCount was modified to return [long] rather than [int].
javadoc edit for SPO

toString() writes out keys correctly for AccessPath.


git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@755 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/inf/Rule.java
@@ -1226,7 +1226,7 @@ abstract public class Rule {
                          * pattern.
                          */
                         
-                        if(minRangeCount == Integer.MAX_VALUE) {
+                        if(minRangeCount == Long.MAX_VALUE) {
                             
                             // range count of the current best choice (computed lazily).
                             minRangeCount = getAccessPath(index,false/*asBound*/).rangeCount();
